+++ test/iterable/iterable_jobs.rb
@@ -1,16 +1,15 @@
 # frozen_string_literal: true
 
 class SimpleIterableJob
-  include Sidekiq::Job
-  include Sidekiq::Job::Iterable
+  include Sidekiq::IterableJob
 
   cattr_accessor :iterated_objects, default: []
   cattr_accessor :on_start_called, default: 0
   cattr_accessor :around_iteration_called, default: 0
   cattr_accessor :on_resume_called, default: 0
-  cattr_accessor :on_shutdown_called, default: 0
+  cattr_accessor :on_stop_called, default: 0
   cattr_accessor :on_complete_called, default: 0
-  cattr_accessor :lifecycle
+  cattr_accessor :context
 
   def on_start
     self.class.on_start_called += 1
@@ -25,8 +24,8 @@ class SimpleIterableJob
     self.class.on_resume_called += 1
   end
 
-  def on_shutdown
-    self.class.on_shutdown_called += 1
+  def on_stop
+    self.class.on_stop_called += 1
   end
 
   def on_complete
@@ -157,13 +156,6 @@ class AbortingIterableJob < ArrayIterableJob
 end
 
 class LongRunningIterableJob < ArrayIterableJob
-  def each_iteration(*)
-    sleep(0.01)
-  end
-end
-
-class LongRunningCustomConfigIterableJob < ArrayIterableJob
-  sidekiq_options iteration: {max_job_runtime: 0.01}
   def each_iteration(*)
     sleep(0.01)
   end